cutters/py-rest/{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = "Code for {{cookiecutter.project_name}}"
authors = ["{{cookiecutter.author}} <{{cookiecutter.email}}>"]

[tool.poetry.dependencies]
python = "3.9.*"
requests = "^2.28"
sanic = "^22.6"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.39"}
alembic = "^1.8.1"
asyncpg = "^0.26.0"
sanic-openapi = "^21.12.0"

[tool.poetry.dev-dependencies]
pytest-mock = "^3.8"
pytest-cov = "^3.0"
autopep8 = "^1.6"
furo = "^2022.6.0"
sphinx = "^5.0"
pytest-asyncio = "^0.19.0"
sanic-testing = "^22.6.0"

[tool.autopep8]
max_line_length = 100
ignore = "E501,W6"  # or ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

[tool.coverage.run]
disable_warnings = ["couldnt-parse"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"